Complete separation of dev calcs and plots
Functions dedicated to the calculation of deviations and the plotting of said deviations have completely been pulled out of plot_funcs.py into dev_calcs.py and dev_plots.py. Minor bug fixed were performed in the process.

Other changes:
- analyze_states.py import_data function has been updated to save and import the dig in times to .csv files after pulling from the original digital_in data format. This significantly speeds up future runs of this code for plot generation or analysis.
